services/auth-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
psycopg2-binary==2.9.9
bcrypt==4.1.1
pyjwt==2.8.0
email-validator==2.1.0
cryptography>=42.0.0